Move validations to utils
